agents/requirements.txt
# Agent Studio requirements with enhanced features
pyyaml>=6.0
python-dotenv>=1.0.0
agno>=0.1.0
loguru>=0.7.0
requests>=2.31.0
openai>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0